% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_content3_efficacy_viz.R
\name{ess_viz}
\alias{ess_viz}
\alias{mwt_viz}
\title{Create Efficacy Viz}
\usage{
ess_viz(
  study_id,
  db_path = system.file("extdata/jazz.sqlite", package = "jazzR")
)

mwt_viz(
  study_id,
  db_path = system.file("extdata/jazz.sqlite", package = "jazzR")
)
}
\arguments{
\item{study_id}{study_id to look up on study_table}

\item{db_path}{sqlite path}
}
\description{
This function returns a ggplot object

This function returns a ggplot object
}
