% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_content1_study_design.R
\name{generate_study_design_table}
\alias{generate_study_design_table}
\title{Generate gt table of patient demographics summary}
\usage{
generate_study_design_table(
  study_id,
  db_path = system.file("extdata/jazz.sqlite", package = "jazzR")
)
}
\arguments{
\item{study_id}{study_id to look up on study_table}

\item{db_path}{sqlite path}
}
\description{
This function calls other content1 functions to calculate sub-tables, joins them, and outputs a gt table.
}
