% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_content1_study_design.R
\name{get_age_summary}
\alias{get_age_summary}
\title{Summarize study subjects age summary}
\usage{
get_age_summary(study_id, db_path)
}
\arguments{
\item{study_id}{study_id to look up on study_table}

\item{db_path}{sqlite path}
}
\description{
Given a study_id, this function returns a min/max/avg summary of age.
}
