% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_content2_efficacy_table.R
\name{get_baseline_ess}
\alias{get_baseline_ess}
\title{Summarize Baseline Mean ESS}
\usage{
get_baseline_ess(study_id, db_path)
}
\arguments{
\item{study_id}{study_id to look up on study_table}

\item{db_path}{sqlite path}
}
\description{
Given a study_id, this function returns mean ESS and
standard deviation by drug. Returns a list of two dataframes, 1 with
re-usable values, 1 with gt-format.
}
