% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_content2_efficacy_table.R
\name{get_change_baseline_placebo_ess}
\alias{get_change_baseline_placebo_ess}
\title{Summarize Mean ESS Change from Baseline and Placebo}
\usage{
get_change_baseline_placebo_ess(study_id, db_path)
}
\arguments{
\item{study_id}{study_id to look up on study_table}

\item{db_path}{sqlite path}
}
\description{
Given a study_id, this function returns the change in mean ESS between
baseline and week 12, as well as week 12 results differences between
drugs and Placebo
}
