% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_content1_study_design.R
\name{get_sex_summary}
\alias{get_sex_summary}
\title{Summarize study subjects sex summary}
\usage{
get_sex_summary(study_id, db_path)
}
\arguments{
\item{study_id}{study_id to look up on study_table}

\item{db_path}{sqlite path}
}
\description{
Given a study_id, this function returns a % summary of sex.
}
