% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_content1_study_design.R
\name{get_study_subjects}
\alias{get_study_subjects}
\title{Summarize Drug Usage by subjects}
\usage{
get_study_subjects(study_id, db_path)
}
\arguments{
\item{study_id}{study_id to look up on study_table}

\item{db_path}{sqlite path}
}
\description{
Given a study_id, this function returns a count summary of drugs assigned to test subjects.
}
