% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_side_bar.R
\name{mod_side_bar_ui}
\alias{mod_side_bar_ui}
\title{side_bar UI Function}
\usage{
mod_side_bar_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
